+++ src/sixel/mod.rs
@@ -1,0 +1,7 @@
+mod detect_support;
+mod sixel_renderer;
+
+pub use {
+    detect_support::detect_sixel_support,
+    sixel_renderer::SixelRenderer,
+};